feat: Windows build fixes and warning cleanup
- Fixed 2777+ compiler warnings (reduced to ~0-2)
- Fixed sign-compare warnings with proper static_cast
- Fixed unused-parameter warnings
- Fixed overloaded-virtual warnings with pragma directives
- Fixed reorder warnings in constructors
- Added build scripts and documentation
- Updated CMakeLists.txt for Windows build
- Added GoogleTest integration
- Improved code quality and compilation compatibility

This commit includes all warning fixes and build improvements for Windows MinGW environment.

+++ include/indicator_utils.h
@@ -63,7 +63,7 @@ inline std::shared_ptr<LineSingle> getDataLine(const std::shared_ptr<LineSeries>
         // DataSeries with OHLCV - use close price at index 4
         // Standard order: DateTime=0, Open=1, High=2, Low=3, Close=4, Volume=5, OpenInterest=6
         return data_source->lines->getline(4);
-    } else if (data_source->lines->size() > default_line) {
+    } else if (data_source->lines->size() > static_cast<size_t>(default_line)) {
         // LineSeries - use specified line (default 0)
         return data_source->lines->getline(default_line);
     }
